removes superset from list of DBs to back up
db called superset is present on production but not on staging.
On production, it has tables that are also present in superset_meta
and table like ab_user is empty. So, it appears to be that it was set up
with initial installation and then superset_meta was setup as the db for
superset